requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
email-validator==2.0.0
kiteconnect==5.0.1

# Flask web app
Flask==3.0.0
gunicorn==21.2.0
